refactor: update imports in auth pages
