+++ src/features/auth/ui/index.ts
@@ -3,6 +3,8 @@ export * from './confirmEmail'
 export * from './createNewPassword'
 export * from './expiredEmail'
 export * from './forgotPassword'
+export * from './privacyPolicy'
 export * from './sentEmail'
 export * from './signIn'
 export * from './signUp'
+export * from './termsOfService'
