--- src/pages/auth/privacy/index.tsx
@@ -1,7 +1,0 @@
-import { PrivacyPolicy } from '@/features/auth/ui/privacyPolicy'
-
-const PrivacyPolicyPage = () => {
-  return <PrivacyPolicy />
-}
-
-export default PrivacyPolicyPage
