--- src/pages/auth/service/index.tsx
@@ -1,7 +1,0 @@
-import { TermsOfService } from '@/features/auth/ui/termsOfService'
-
-const TermsOfServicePage = () => {
-  return <TermsOfService />
-}
-
-export default TermsOfServicePage
